ScheduleTask/requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.12.2
idna==2.10
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
